requirements.txt
slack-bolt==1.18.1
flask==3.0.0
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
